Updated solution.hide.js
The solution as it was did not take into account the possibility of both numbers being negative. The previous code would return 0 if both numbers are negative, which is not the solution (e.g., -2 * -7 = 14). To address this, I added a conditional that returns the absolute value of "aux" when both numbers are negative, resulting in the correct product.